Bring in release scripts from mantisbt-tools
We had our build tools split between MantisBT main repo and mantisbt-tools repo.
It is a better approach to have the scripts in the main repo, so each MantisBT branch
can have the corresponding version of the build scripts that works with it.  Also it makes
it easier to keep the build scripts up-to-date as the MantisBT core code changes.